Update ru_ru.json
Most likely, this is slang, implying that there's no need for a second translation since BardinTheDwarf's RP version already received a like. However, I double-checked the information about some creature names and will make new adjustments. The subtitles were taken exclusively from the Russian localization of the game itself. On top of that, there are no subtitles for the Incinerator.